requirements.txt
Django~=2.2
djangorestframework~=3.12.2